fix: Return Numeric type for aggregate functions (COUNT, SUM, AVG)
Changes aggregate functions to return SqlValue::Numeric instead of
SqlValue::Integer for consistency with SQLLogicTest expectations.

Changes:
- Updated COUNT to return Numeric in grouping.rs
- Updated COUNT, SUM, AVG to return Numeric in window aggregates
- Fixed COUNT(*) fast paths to return Numeric
- Updated test expectations to match new return types

This ensures aggregate results format with .000 decimal suffix
as expected by SQLLogicTest, fixing formatting inconsistencies.

Fixes #1171

+++ crates/executor/src/select/grouping.rs
@@ -187,7 +187,7 @@ impl AggregateAccumulator {
 
     pub(super) fn finalize(&self) -> types::SqlValue {
         match self {
-            AggregateAccumulator::Count { count, .. } => types::SqlValue::Integer(*count),
+            AggregateAccumulator::Count { count, .. } => types::SqlValue::Numeric(*count as f64),
             AggregateAccumulator::Sum { sum, .. } => sum.clone(),
             AggregateAccumulator::Avg { sum, count, .. } => {
                 if *count == 0 {
